Completes analyzeRabiAmp().
git-svn-id: https://qubit.bbn.com/svn/MQCO/trunk@271 005577ff-ac90-44ea-9eee-c3a23b7cba77

+++ common/src/+expManager/@pulseCalibration/pulseCalibrationDo.m
@@ -91,7 +91,7 @@ if ExpParams.DoPi2Cal
     fprintf('Found Q offset: %.3f\n', q_offset);
     
     % update pulseParams
-    obj.pulseParams.pi2Amp = Y90Amp;
+    obj.pulseParams.pi2Amp = int32(Y90Amp);
     obj.pulseParams.i_offset = i_offset;
     obj.pulseParams.q_offset = q_offset;
     % update T matrix with ratio X90Amp/Y90Amp
@@ -116,7 +116,7 @@ if ExpParams.DoPiCal
     fprintf('Found X180Amp: %.0f\n', X180Amp);
     
     % update pulseParams
-    obj.pulseParams.piAmp = X180Amp;
+    obj.pulseParams.piAmp = int32(X180Amp);
 end
 
 %% DRAG calibration
